requirements.txt
python-telegram-bot==21.3
yt-dlp
ffmpeg-python